server/requirements.txt
asgiref==3.9.1
bcrypt==4.3.0
certifi==2025.8.3
charset-normalizer==3.4.2
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
dnspython==2.7.0
idna==3.10
mailjet-rest==1.5.1
PyJWT==2.10.1
pymongo==4.14.0
python-dotenv==1.1.1
requests==2.32.4
sqlparse==0.5.3
urllib3==2.5.0
